+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to [@bpmn-io/cm-theme](https://github.com/bpmn-io/cm-theme) 
 ___Note:__ Yet to be released changes appear here._
 
 - `FIX`: restore feelers syntax highlighting ([#16](https://github.com/bpmn-io/cm-theme/pull/16))
+- `FIX`: restore lint range squiggle underlines ([#16](https://github.com/bpmn-io/cm-theme/pull/16))
 
 ## 0.2.0